增加，查询，删除

+++ Spring-boot-demo/src/main/java/com/example/demo/controller/UserController.java
@@ -1,11 +1,17 @@
 package com.example.demo.controller;
 
+import com.example.demo.exception.UserNotFoundException;
 import com.example.demo.model.User;
+import com.example.demo.util.Error;
 import com.example.demo.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/user")
@@ -59,6 +65,14 @@ public class UserController {
 //        return list;
 //    }
 
+    @ExceptionHandler(UserNotFoundException.class)
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    public Error userNotFound(UserNotFoundException e) {
+        long uid = e.getUid();
+        Error error = new Error(10001,"user " + uid + " not found");
+        return error;
+    }
+
     @RequestMapping(value = "/list",method = {RequestMethod.GET})
     public List<User> getUsers() {
         List list = userService.findAllUser();
@@ -69,13 +83,34 @@ public class UserController {
     public User getUser(@PathVariable Integer id) {
 
         User user = userService.findUser(id);
+        if (user == null) {
+            throw new UserNotFoundException(id);
+        }
         return user;
     }
 
     @RequestMapping(value = "/{id}",method = {RequestMethod.DELETE})
-    public void deleteUser(@PathVariable Integer id) {
+    public Map deleteUser(@PathVariable Integer id) {
 
-        userService.deleteUser(id);
+        User user = userService.findUser(id);
+        if (user == null) {
+            throw new UserNotFoundException(id);
+        } else {
+            userService.deleteUser(id);
+        }
+        Map map = new HashMap();
+        map.put("result","success");
+        map.put("uid",id);
+        return map;
+    }
+
+    @RequestMapping(value = "/add",method = {RequestMethod.POST})
+    public User addUser(User user) {
+        userService.addUserSelective(user);
+
+        User theUser = userService.findUser(user.uid);
+
+        return theUser;
     }
 
     @ModelAttribute
@@ -84,7 +119,7 @@ public class UserController {
 //        User user = userService.findUser(id);
 
 //        User theUser = userService.updateUser(user);
-        
+
         return null;
     }
 }
